fix: move gen_nonsense_data to bartz.testing
