+++ src/bartz/testing/_nonsense.py
@@ -1,4 +1,4 @@
-# bartz/src/bartz/testing/__init__.py
+# bartz/src/bartz/testing/_nonsense.py
 #
 # Copyright (c) 2026, The Bartz Contributors
 #
@@ -22,9 +22,28 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-"""Testing utilities."""
 
-# ruff: noqa: F401
+"""Define `gen_nonsense_data`, a cheap synthetic data generator for benchmarks."""
 
-from bartz.testing._dgp import DGP, Params, gen_data, gen_data_from_params, gen_params
-from bartz.testing._nonsense import gen_nonsense_data
+from functools import partial
+
+from jax import jit, vmap
+from jax import numpy as jnp
+from jaxtyping import Array, Float32, UInt8
+
+
+@partial(jit, static_argnums=(0, 1, 2))
+def gen_nonsense_data(
+    p: int, n: int, k: int | None
+) -> tuple[
+    UInt8[Array, '{p} {n}'],
+    Float32[Array, ' {n}'] | Float32[Array, '{k} {n}'],
+    UInt8[Array, ' {p}'],
+]:
+    """Generate pretty nonsensical data."""
+    X = jnp.arange(p * n, dtype=jnp.uint8).reshape(p, n)
+    X = vmap(jnp.roll)(X, jnp.arange(p))
+    max_split = jnp.full(p, 255, jnp.uint8)
+    shift = 0 if k is None else jnp.linspace(0, 2 * jnp.pi, k, endpoint=False)[:, None]
+    y = jnp.cos(jnp.linspace(0, 2 * jnp.pi / 32 * n, n) + shift)
+    return X, y, max_split
